--- .env
@@ -1,1 +1,0 @@
-SERVER_PORT="3000";
