--- README.md
@@ -1,6 +1,0 @@
-1- missing readme.
-2- project name no special chars / captial letters.
-3- port from env var.
-4- inner function shouldn't be exposed to the HTTP protocol
-5- fix issues from the comments.
-6- try convert the app to typescript.
