+++ module/db.js
@@ -25,4 +25,4 @@ let userSchema = new mongoose.Schema({
   },
 });
 
-module.exports = mongoose.model("user", userSchema);
+module.exports = mongoose.model("use", userSchema);
