--- package.json
@@ -1,23 +1,0 @@
-{
-  "name": "res_test",
-  "version": "1.0.0",
-  "description": "1- missing readme.\r 2- project name no special chars / captial letters.\r 3- port from env var.\r 4- inner function shouldn't be exposed to the HTTP protocol\r 5- fix issues from the comments.\r 6- try convert the app to typescript.",
-  "main": "index.js",
-  "dependencies": {
-    "body-parser": "^1.20.1",
-    "dotenv": "^16.0.3",
-    "express": "^4.18.2",
-    "mongoose": "^6.7.1",
-    "node-fetch": "^3.2.10",
-    "typescript": "^4.8.4"
-  },
-  "devDependencies": {
-    "@types/express": "^4.17.14",
-    "@types/node": "^18.11.9"
-  },
-  "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
-  },
-  "author": "",
-  "license": "ISC"
-}
