+++ routes/users.js
@@ -1,174 +1,61 @@
-"use strict";
-var __awaiter = (this && this.__awaiter) || function (thisArg, _arguments, P, generator) {
-    function adopt(value) { return value instanceof P ? value : new P(function (resolve) { resolve(value); }); }
-    return new (P || (P = Promise))(function (resolve, reject) {
-        function fulfilled(value) { try { step(generator.next(value)); } catch (e) { reject(e); } }
-        function rejected(value) { try { step(generator["throw"](value)); } catch (e) { reject(e); } }
-        function step(result) { result.done ? resolve(result.value) : adopt(result.value).then(fulfilled, rejected); }
-        step((generator = generator.apply(thisArg, _arguments || [])).next());
-    });
-};
-var __generator = (this && this.__generator) || function (thisArg, body) {
-    var _ = { label: 0, sent: function() { if (t[0] & 1) throw t[1]; return t[1]; }, trys: [], ops: [] }, f, y, t, g;
-    return g = { next: verb(0), "throw": verb(1), "return": verb(2) }, typeof Symbol === "function" && (g[Symbol.iterator] = function() { return this; }), g;
-    function verb(n) { return function (v) { return step([n, v]); }; }
-    function step(op) {
-        if (f) throw new TypeError("Generator is already executing.");
-        while (_) try {
-            if (f = 1, y && (t = op[0] & 2 ? y["return"] : op[0] ? y["throw"] || ((t = y["return"]) && t.call(y), 0) : y.next) && !(t = t.call(y, op[1])).done) return t;
-            if (y = 0, t) op = [op[0] & 2, t.value];
-            switch (op[0]) {
-                case 0: case 1: t = op; break;
-                case 4: _.label++; return { value: op[1], done: false };
-                case 5: _.label++; y = op[1]; op = [0]; continue;
-                case 7: op = _.ops.pop(); _.trys.pop(); continue;
-                default:
-                    if (!(t = _.trys, t = t.length > 0 && t[t.length - 1]) && (op[0] === 6 || op[0] === 2)) { _ = 0; continue; }
-                    if (op[0] === 3 && (!t || (op[1] > t[0] && op[1] < t[3]))) { _.label = op[1]; break; }
-                    if (op[0] === 6 && _.label < t[1]) { _.label = t[1]; t = op; break; }
-                    if (t && _.label < t[2]) { _.label = t[2]; _.ops.push(op); break; }
-                    if (t[2]) _.ops.pop();
-                    _.trys.pop(); continue;
-            }
-            op = body.call(thisArg, _);
-        } catch (e) { op = [6, e]; y = 0; } finally { f = t = 0; }
-        if (op[0] & 5) throw op[1]; return { value: op[0] ? op[1] : void 0, done: true };
+const express = require("express");
+const router = express.Router();
+const db = require("./../module/db");
+
+router.get("/:id", getUsers, (req, res) => {
+  res.json(res.user);
+});
+
+router.post("/", async (req, res) => {
+  let Db = new db({
+    firstName: req.body.firstName,
+    lastName: req.body.lastName,
+    age: req.body.age,
+  });
+  try {
+    Db = await Db.save();
+    res.status(201).send(`User ${Db.firstName} Created Details :` + Db);
+  } catch (err) {
+    res.send("error : " + err);
+  }
+});
+router.patch("/:id", getUsers, async (req, res) => {
+  if (req.body.firstName != null) {
+    res.user.firstName = req.body.firstName;
+  }
+  if (req.body.lastName != null) {
+    res.user.lastName = req.body.lastName;
+  }
+  if (req.body.age != null) {
+    res.user.age = req.body.age;
+  }
+  try {
+    let updateUser = await res.user.save();
+    res.send(`User ${updateUser.firstName} Update Details :` + updateUser);
+  } catch (err) {
+    res.status(400).json({ message: "Bad Request" });
+  }
+});
+router.delete("/:id", getUsers, async (req, res) => {
+  try {
+    await res.user.remove();
+    res.json({ message: "Deleted" });
+  } catch (err) {
+    res.status(500).json({ message: "Internal Server Error" });
+  }
+});
+
+async function getUsers(req, res, next) {
+  let user;
+  try {
+    user = await db.findById(req.params.id);
+    if (user == null) {
+      return res.status(404).json({ message: "Not Found" });
     }
-};
-exports.__esModule = true;
-var express_1 = require("express");
-var router = express_1["default"].Router();
-var db = require("./../module/db");
-router.get("/:id", findUserById);
-// inner function shouldn't be exposed to the HTTP protocol
-// getUsetById(id: string) => user
-function findUserById(req, res) {
-    return __awaiter(this, void 0, void 0, function () {
-        var id;
-        return __generator(this, function (_a) {
-            switch (_a.label) {
-                case 0:
-                    id = req.params.id;
-                    return [4 /*yield*/, db.findById(id)
-                            .then(function (user) {
-                            if (user == null) {
-                                res.status(404).json({ message: "Not Found user" });
-                            }
-                            res.status(200).send("User ".concat(user.firstName) + user);
-                        })["catch"](function () {
-                            res.status(500).send({ message: "Canot find this id=".concat(id) });
-                        })];
-                case 1:
-                    _a.sent();
-                    return [2 /*return*/];
-            }
-        });
-    });
-}
-router.post("/", function (req, res) { return __awaiter(void 0, void 0, void 0, function () {
-    var _a, firstName, lastName, age, Db, user, err_1;
-    return __generator(this, function (_b) {
-        switch (_b.label) {
-            case 0:
-                _a = req.body, firstName = _a.firstName, lastName = _a.lastName, age = _a.age;
-                Db = new db({
-                    firstName: firstName,
-                    lastName: lastName,
-                    age: age
-                });
-                user = {
-                    firstName: firstName,
-                    lastName: lastName,
-                    age: age
-                };
-                _b.label = 1;
-            case 1:
-                _b.trys.push([1, 3, , 4]);
-                return [4 /*yield*/, Db.save()];
-            case 2:
-                Db = _b.sent();
-                res.status(201).send("User ".concat(user.firstName, " Created Details :") + Db);
-                return [3 /*break*/, 4];
-            case 3:
-                err_1 = _b.sent();
-                res.send("error : " + err_1);
-                return [3 /*break*/, 4];
-            case 4: return [2 /*return*/];
-        }
-    });
-}); });
-router.patch("/:id", getUsers, function (req, res) { return __awaiter(void 0, void 0, void 0, function () {
-    var updateUser, err_2;
-    return __generator(this, function (_a) {
-        switch (_a.label) {
-            case 0:
-                if (req.body.firstName != null) {
-                    res.user.firstName = req.body.firstName;
-                }
-                if (req.body.lastName != null) {
-                    res.user.lastName = req.body.lastName;
-                }
-                if (req.body.age != null) {
-                    res.user.age = req.body.age;
-                }
-                _a.label = 1;
-            case 1:
-                _a.trys.push([1, 3, , 4]);
-                return [4 /*yield*/, res.user.save()];
-            case 2:
-                updateUser = _a.sent();
-                res.send("User ".concat(updateUser.firstName, " Update Details :") + updateUser);
-                return [3 /*break*/, 4];
-            case 3:
-                err_2 = _a.sent();
-                res.status(400).json({ message: "Bad Request" });
-                return [3 /*break*/, 4];
-            case 4: return [2 /*return*/];
-        }
-    });
-}); });
-router["delete"]("/:id", getUsers, function (req, res) { return __awaiter(void 0, void 0, void 0, function () {
-    var err_3;
-    return __generator(this, function (_a) {
-        switch (_a.label) {
-            case 0:
-                _a.trys.push([0, 2, , 3]);
-                return [4 /*yield*/, res.user.remove()];
-            case 1:
-                _a.sent();
-                res.json({ message: "Deleted" });
-                return [3 /*break*/, 3];
-            case 2:
-                err_3 = _a.sent();
-                res.status(500).json({ message: "Internal Server Error" });
-                return [3 /*break*/, 3];
-            case 3: return [2 /*return*/];
-        }
-    });
-}); });
-function getUsers(req, res, next) {
-    return __awaiter(this, void 0, void 0, function () {
-        var user, err_4;
-        return __generator(this, function (_a) {
-            switch (_a.label) {
-                case 0:
-                    _a.trys.push([0, 2, , 3]);
-                    return [4 /*yield*/, db.findById(req.params.id)];
-                case 1:
-                    user = _a.sent();
-                    if (user == null) {
-                        return [2 /*return*/, res.status(404).json({ message: "Not Found" })];
-                    }
-                    return [3 /*break*/, 3];
-                case 2:
-                    err_4 = _a.sent();
-                    return [2 /*return*/, res.status(500).json({ message: "Internal Server Error" })];
-                case 3:
-                    res.user = user;
-                    next();
-                    return [2 /*return*/];
-            }
-        });
-    });
+  } catch (err) {
+    return res.status(500).json({ message: "Internal Server Error" });
+  }
+  res.user = user;
+  next();
 }
 module.exports = router;
